[RENAME] AuthorizationViewModel로 권한 확인하는 viewModel로 변경 및 통합
