+++ Sofa/Sources/View/Album/SubViews/Authorization/CameraAuthorization.swift
@@ -5,7 +5,6 @@
 //  Created by geonhyeong on 2022/06/16.
 //
 
-import SwiftUI
 import AVFoundation
 
 struct CameraAuthorization {
